refactor: move useFetch hook to ContentViewSwitch directory
