+++ frontend/src/ui/share/ContentViewSwitch/useFetch.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 
 export const useFetch = <T extends {}>(fetchFunc: () => Promise<T>, dependencies: any[]) => {
   const [result, setResult] = useState<T | undefined>(undefined);
-  const [error, setError] = useState('');
+  const [error, setError] = useState();
   const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
